Move _StderrLogger to logger/stderr.dart
